Move isBot() method to GameDescription
* Also determined that GameDescription returned by various table models
  cannot be null, so removed assocaited null checks.
