+++ game-core/src/main/java/games/strategy/engine/lobby/client/ui/LobbyGameTableModel.java
@@ -149,7 +149,7 @@ class LobbyGameTableModel extends AbstractTableModel {
       case P:
         return (description.getPassworded() ? "*" : "");
       case B:
-        return (description.getBotSupportEmail() != null && description.getBotSupportEmail().length() > 0 ? "-" : "");
+        return description.isBot() ? "-" : "";
       case GV:
         return description.getGameVersion();
       case EV:
